Rename .h to .hpp
