+++ Logging/LoggingModule.hpp
