+++ Logging/LoggingView.hpp
